factor out futil candidates
