+++ src/utils/fields.js
@@ -1,6 +1,5 @@
 let _ = require('lodash/fp')
-
-let maybeAppend = (suffix, str) => _.endsWith(suffix, str) ? str : str + suffix
+let { maybeAppend } = require('./futil')
 
 let dot = x => x ? `.${x}` : ''
 
@@ -8,6 +7,5 @@ let path = (schema, field) =>
   dot(_.get(['fields', field, 'elasticsearch', 'notAnalyzedField'], schema))
 
 module.exports = {
-  maybeAppend,
   getField: (schema, field) => maybeAppend(path(schema, field), field),
 }
